+++ src/lib.rs
@@ -17,8 +17,10 @@ pub use shulkerbox;
 
 pub mod base;
 pub mod lexical;
+pub mod semantic;
 pub mod syntax;
 pub mod transpile;
+pub mod util;
 
 #[cfg(feature = "serde")]
 pub(crate) mod serde;
@@ -105,6 +107,8 @@ pub fn parse(
             "An error occurred while parsing the source code.",
         ));
     }
+
+    program.analyze_semantics(handler)?;
 
     Ok(program)
 }