+++ src/transpile/error.rs
@@ -8,9 +8,9 @@ use itertools::Itertools;
 use crate::{
     base::{
         log::{Message, Severity, SourceCodeDisplay},
-        source_file::{SourceElement, Span},
+        source_file::Span,
     },
-    syntax::syntax_tree::expression::Expression,
+    semantic::error::{ConflictingFunctionNames, InvalidFunctionArguments, UnexpectedExpression},
 };
 
 use super::FunctionData;
@@ -29,6 +29,8 @@ pub enum TranspileError {
     LuaRuntimeError(#[from] LuaRuntimeError),
     #[error(transparent)]
     ConflictingFunctionNames(#[from] ConflictingFunctionNames),
+    #[error(transparent)]
+    InvalidFunctionArguments(#[from] InvalidFunctionArguments),
 }
 
 /// The result of a transpilation operation.
@@ -144,52 +146,3 @@ impl LuaRuntimeError {
         }
     }
 }
-
-/// An error that occurs when a function declaration is missing.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct UnexpectedExpression(pub Expression);
-
-impl Display for UnexpectedExpression {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            Message::new(Severity::Error, "encountered unexpected expression")
-        )?;
-
-        write!(
-            f,
-            "\n{}",
-            SourceCodeDisplay::new(&self.0.span(), Option::<u8>::None)
-        )
-    }
-}
-
-impl std::error::Error for UnexpectedExpression {}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ConflictingFunctionNames {
-    pub definition: Span,
-    pub name: String,
-}
-
-impl Display for ConflictingFunctionNames {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            Message::new(
-                Severity::Error,
-                format!("the following function declaration conflicts with an existing function with name `{}`", self.name)
-            )
-        )?;
-
-        write!(
-            f,
-            "\n{}",
-            SourceCodeDisplay::new(&self.definition, Option::<u8>::None)
-        )
-    }
-}
-
-impl std::error::Error for ConflictingFunctionNames {}