+++ src/transpile/util.rs
@@ -1,3 +1,8 @@
+//! Utility methods for transpiling
+
+#[cfg(feature = "shulkerbox")]
+use chksum_md5 as md5;
+
 fn normalize_program_identifier<S>(identifier: S) -> String
 where
     S: AsRef<str>,
@@ -19,6 +24,8 @@ where
         .join("/")
 }
 
+/// Calculate the identifier to import the function based on the current identifier and the import path
+#[must_use]
 pub fn calculate_import_identifier<S, T>(current_identifier: S, import_path: T) -> String
 where
     S: AsRef<str>,
@@ -30,5 +37,27 @@ where
         let mut identifier_elements = current_identifier.as_ref().split('/').collect::<Vec<_>>();
         identifier_elements.pop();
         normalize_program_identifier(identifier_elements.join("/") + "/" + import_path.as_ref())
+    }
+}
+
+/// Transforms an identifier to a macro name that only contains `a-zA-Z0-9_`.
+#[cfg(feature = "shulkerbox")]
+#[must_use]
+pub fn identifier_to_macro(ident: &str) -> std::borrow::Cow<str> {
+    if ident.contains("__")
+        || ident
+            .chars()
+            .any(|c| !(c == '_' && c.is_ascii_alphanumeric()))
+    {
+        let new_ident = ident
+            .chars()
+            .filter(|c| *c == '_' || c.is_ascii_alphanumeric())
+            .collect::<String>();
+
+        let chksum = md5::hash(ident).to_hex_lowercase();
+
+        std::borrow::Cow::Owned(new_ident + "__" + &chksum[..8])
+    } else {
+        std::borrow::Cow::Borrowed(ident)
     }
 }