+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Macro strings
+- Function parameters/arguments
+
 ### Changed
 
 - Option to deduplicate source files during serialization when using `SerdeWrapper`